+++ CHANGELOG.md
@@ -7,4 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* basic routing CLI
+
 [Unreleased]: https://github.com/escaped/routor/tree/master